requirements.txt
sentence-transformers==2.7.0
chromadb==0.5.3
google-generativeai==0.7.2
python-dotenv==1.0.1
PyYAML==6.0.2
tqdm==4.66.4
numpy==1.26.4
